[Create] 몬스터 공격 애니메이션

+++ Assets/WorkSpace/LJH/Scriptable/Monster/MonsterData.cs
@@ -9,7 +9,7 @@ public class MonsterData : ScriptableObject
     public int id;
     public int hp;
     public int attack;
-    public int range;
+    public float range;
     public float cooldown;
 
     //추후 아이템 클래스 생성시 그것으로 교체
